bgpd: fix SID structure advertised in SRv6 SID Structure TLV
In the current implementation, when bgpd creates an
SRv6 SID Structure Sub-Sub-TLV message to advertise a SID,
it ignores the bits length parameters configured in the SRv6 locator.
Instead, it always advertises the following SID stucture (the values
are hardcoded):
* block-len = 40
* node-len = 24
* function-len = 16
* argument-len = 0
* transposition-len = 16
* transposition-offset = 64

Consequently, if an user creates a SRv6 locator with a different SID
structure, bgpd will advertise a wrong SID structure to its peers.

This commit changes bgpd to advertise the SID structure of the SRv6
locator.

Signed-off-by: Carmine Scarpitta <carmine.scarpitta@uniroma2.it>