doc: Add doc for `source-address` command
The `source-address` command allows users to set the source address used
for the outer encapsulating IPv6 header. When you configure an address
with the `source-address` command, the source address is stored in the
`struct zebra_srv6` data structure owned by zebra and installed in the
data plane.

In this commit, we add the documentation for the `source-address`
command.

Signed-off-by: Carmine Scarpitta <carmine.scarpitta@uniroma2.it>