zebra: Allow resolving Generic Netlink family
Generic Netlink supports the dynamic allocation of communication
channels. Communication channels are uniquely identified by channel
numbers which are dynamically allocated by a Generic Netlink controller.
The controller listens on a fixed communication channel (channel number
0x10). Clients who want to use a service query the controller to see if
the service exists and to determine the correct channel number.

In this commit, we add the base functionality to get the channel number
assigned to a specific service. In particular, we add a function
`genl_resolve_family()` that takes the service name (called family in
the Generic Netlink terminology) as input and queries the Generic
Netlink controller to get the channel number that corresponds to the
service.

Signed-off-by: Carmine Scarpitta <carmine.scarpitta@uniroma2.it>