zebra: Allow resolving Generic Netlink SEG6 family
Generic Netlink supports the dynamic allocation of communication
channels. Communication channels are uniquely identified by channel
numbers which are dynamically allocated by a Generic Netlink controller.
The controller listens on a fixed communication channel (channel number
0x10). Clients who want to use a service query the controller to see if
the service exists and to determine the correct channel number.

Generic Netlink family resolution has been introduced in a previous
commit. In this commit, we extend zebra to discover the channel number
associated with the `SEG6` family. zebra queries the Generic Netlink
controller to get the channel number that corresponds to the
`SEG6` family. Then, it uses `netlink_seg6_genl_parse_family()` to
parse the reply receive from the controller and extract the `SEG6`
channel number.

Signed-off-by: Carmine Scarpitta <carmine.scarpitta@uniroma2.it>