zebra: Add CLI command to `show srv6 information`
This commit adds a new CLI command `show srv6 information` that is used
to print the SRv6 information. Currently, the SRv6 information includes
the SRv6 encapsulation source address and the SRv6 locators configured
in zebra.

Below is an example of the output of the `show srv6 information`
command:

```
router# show segment-routing srv6 json
{
  "encapsulation":{
    "sourceAddress":"1111:2222:3333:4444::8888"
  },
  "locators":[
    {
      "name":"loc1",
      "prefix":"fc00:0:1::/64",
      "functionBitsLength":16,
      "statusUp":true,
      "chunks":[
        {
          "prefix":"fc00:0:1::/64",
          "proto":"bgp"
        }
      ]
    }
  ]
}
```

Signed-off-by: Carmine Scarpitta <carmine.scarpitta@uniroma2.it>